Debugs remove lenses
Whenever a lens is removed, it's place in the stack is set to "undefined". When
trying to draw the beam and lenses, we first build a second stack similar to
the normal stack but without the undefined lenses. This second stack is then used
for drawing.

When adding a lens, we first look if some ids are free (set to undefined), and use
those before increasing the size of the lens stack.